chore: substitui header hardcoded por componente.
