+++ src/screens/SignUp.jsx
@@ -1,26 +1,13 @@
 import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import CustomInput from "../components/CustomInput.jsx";
+import RegistrationHeader from "../components/RegistrationHeader.jsx";
 
 function SignUp() {
 
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
 
-  const headerStyle = {
-    color: "#06B3C4",
-    textAlign: "center",
-    fontSize: "20px",
-    fontWeight: "600",
-  };
-
-  const subtitleStyle = {
-    color: "#777777",
-    textAlign: "center",
-    fontSize: "16px",
-    fontWeight: "600",
-    margin: "16px 80px",
-  };
 
   const noteStyle = { fontSize: "13px" };
 
@@ -30,7 +17,7 @@ function SignUp() {
     padding: "16px 0",
   };
 
-  const headerText = "Crie sua conta";
+  const titleText = "Crie sua conta";
   const subtitleText = "Preencha os dados abaixo para se cadastrar no aplicativo.";
   const emailNote = "A confirmação e acompanhamento do seu pedido serão enviados ao e-mail cadastrado.";
   const passwordNote = "Use oito ou mais caracteres com uma combinação de letras, números e símbolos para uma senha mais segura.";
@@ -55,10 +42,12 @@ function SignUp() {
         <section className="mt-5">
           <div className="columns">
             <div className="column">
-              <h1 className="subtitle" style={headerStyle}>
-                {headerText}
-              </h1>
-              <h2 style={subtitleStyle}>{subtitleText}</h2>
+
+              <RegistrationHeader
+                title={titleText}
+                subtitle={subtitleText}
+              />
+
               <form onSubmit={handleSubmit}>
                 <CustomInput
                   label="Insira seu e-mail"
